pyproject.toml
[tool.poetry]
name = "repl_python3_Discord-BOT"
version = "0.1.0"
description = ""
authors = ["ElMoha943 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.17.1"
flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
